Revert moving status.cc
